pyproject.toml
[project]
name = "dk_results"
version = "0.1.0"
description = ""
requires-python = ">=3.11,<3.12"

dependencies = [
    "google-api-python-client",
    "requests",
    "pulp",
    "rookiepy",
    "pyyaml",
    "dotenv",
    "discord",
    "dfs-common",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "ty",
]

[tool.uv]
package = true
default-groups = ["dev"]

[tool.uv.sources]
dfs-common = { path = "../dfs_common", editable = true }

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]

[tool.ty]

[tool.ty.src]
exclude = ["tests"]
